+++ MKAgent/src/main/java/org/AIandGames/mancalabot/Enums/TerminalState.java
@@ -1,10 +1,10 @@
 package org.AIandGames.mancalabot.Enums;
 
 public enum TerminalState {
-    WIN_TERMINAL(1),
-    LOSE_TERMINAL(-1),
-    NON_TERMINAL(0);
+    WIN_TERMINAL((byte) 1),
+    LOSE_TERMINAL((byte) -1),
+    NON_TERMINAL((byte) 0);
 
-    TerminalState(final int terminalState) {
+    TerminalState(final byte terminalState) {
     }
 }
